Refactor builder.py to be more like os.path
